optimized navbar for bundle size
