+++ src/components/Layout/Layout/Layout.tsx
@@ -1,8 +1,11 @@
 import React from 'react'
 import { Box, Stack } from '@chakra-ui/react'
+import Navbar from '../Navbar/Navbar'
 
 const Footer = React.lazy(() => import('@/components/Layout/Footer/Footer'))
-const Navbar = React.lazy(() => import('@/components/Layout/Navbar/Navbar'))
+const NetworkErrorNotification = React.lazy(
+  () => import('@/components/Layout/Network/NetworkErrorNotification'),
+)
 const GoogleAnalyticsScripts = React.lazy(
   () => import('@/components/Layout/Layout/GoogleAnalyticsScripts'),
 )
@@ -17,6 +20,7 @@ const Layout = ({
   return (
     <>
       <GoogleAnalyticsScripts />
+
       <Stack justify="space-between" minH="100vh" spacing={0}>
         <Navbar />
         <Box as="main" pt={20}>
@@ -24,6 +28,8 @@ const Layout = ({
         </Box>
         {!noFooter && <Footer />}
       </Stack>
+
+      <NetworkErrorNotification />
     </>
   )
 }
